✨ add: alias for routes, removing styled components, adding sass, rename jsx to tsx and adapting components and pages signIn and signUp
