+++ package.json
@@ -10,6 +10,7 @@
     "normalize.css": "^8.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "sass": "^1.49.9",
     "styled-components": "^5.3.5"
   },
   "scripts": {
