+++ src/components/auth/FooterForm.tsx
@@ -1,4 +1,4 @@
-import { Link } from "react-router-dom"
+import Link from "next/link"
 
 export const FooterForm = ({title, route = "/", textLInk}) => {
     return (
@@ -7,7 +7,9 @@ export const FooterForm = ({title, route = "/", textLInk}) => {
                 <h3 className="footer-auth__title">
                     {title}
                 </h3>
-                <Link className="footer-auth__link"to={route}>{textLInk}</Link>
+                <Link href={route}>
+                    <a className="footer-auth__link" >{textLInk}</a>
+                </Link>
             </div>
             <p className={[route === '/sign-up' && 'footer-auth__terms--mt', 'footer-auth__terms'].join(' ')}>
                 Términos y condiciones de uso
