+++ src/components/auth/FormAuth/GroupInputsReggister1.tsx
@@ -1,4 +1,3 @@
-
 export const GroupInputsReggister1 = ({changeInputValues, inputValues}) => {
     return (
         <>
