+++ src/components/auth/FormAuth/GroupInputsReggister2.tsx
