+++ src/pages/Auth/SignUp.tsx
@@ -1,11 +1,12 @@
-import { ReactComponent as Logo } from "assets/svgs/Logo.svg";
-import { FormAuth } from "components/FormAuth/FormAuth";
-import { FooterForm } from "./components/FooterForm";
-export const SignUp = () => {
+/* import { ReactComponent as Logo } from "assets/svgs/Logo.svg"; */
+import { FormAuth } from "@components/auth/FormAuth";
+
+import { FooterForm } from "@components/auth/FooterForm";
+const SignUp = () => {
 
     return (
         <div className="container-global content-page-auth">
-            <Logo className="content-page-auth__icon" />
+            {/* <Logo className="content-page-auth__icon" /> */}
             <h1 className="content-page-auth__title">
                 Crea una cuenta
             </h1>
@@ -15,9 +16,10 @@ export const SignUp = () => {
             <FooterForm
                 title="¿Ya tienes una cuenta?"
                 textLInk="¡Ingresa ahora!"
-                route="/sign-in"
+                route="/Auth/SignIn"
             /> 
             
         </div>
     )
 }
+export default SignUp
