+++ src/pages/_app.tsx
@@ -1,5 +1,6 @@
 import "normalize.css";
 import "animate.css";
+import "../styles/main.scss";
 function MyApp ({ Component, pageProps }) {
   return (
     <Component {...pageProps} />
